Adiciona dialog de confirmação de exclusão

+++ src/app/courses/containers/courses/courses.component.ts
@@ -7,6 +7,7 @@ import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/err
 
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -44,25 +45,39 @@ export class CoursesComponent implements OnInit {
   ngOnInit(): void {}
 
   onAdd() {
+    this.refresh();
     this.router.navigate(['new'], { relativeTo: this.route });
+    this.refresh();
   }
 
   onEdit(course: Course) {
     this.refresh();
-    this.router.navigate(['edit', course._id], { relativeTo: this.route });
+    this.router
+      .navigate(['edit', course._id], { relativeTo: this.route })
+      .then(() => this.refresh());
+
+    this.refresh();
   }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe({
-      next: () => {
-        this.refresh();
-        this.snackBar.open('Curso removido com sucesso!', 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center',
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe({
+          next: () => {
+            this.refresh();
+            this.snackBar.open('Curso removido com sucesso!', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
+          },
+          error: () => this.onError('Erro ao tentar remover curso.'),
         });
-      },
-      error: () => this.onError('Erro ao tentar remover curso.'),
+      }
     });
   }
 }
